refactor secret getter constructor with defaults
We were creating the secret getter with the same cache size and ttl
in a bunch of different places with the same values.
