+++ pkg/interceptors/gitlab/gitlab_test.go
@@ -19,7 +19,6 @@ package gitlab
 import (
 	"net/http"
 	"testing"
-	"time"
 
 	triggersv1 "github.com/tektoncd/triggers/pkg/apis/triggers/v1beta1"
 	"github.com/tektoncd/triggers/pkg/interceptors"
@@ -121,7 +120,7 @@ func TestInterceptor_ExecuteTrigger_ShouldContinue(t *testing.T) {
 				ctx, clientset = fakekubeclient.With(ctx, tt.secret)
 			}
 			w := &Interceptor{
-				SecretGetter: interceptors.NewKubeClientSecretGetter(clientset.CoreV1(), 1024, 5*time.Second),
+				SecretGetter: interceptors.DefaultSecretGetter(clientset.CoreV1()),
 			}
 			res := w.Process(ctx, req)
 			if !res.Continue {
@@ -272,7 +271,7 @@ func TestInterceptor_ExecuteTrigger_ShouldNotContinue(t *testing.T) {
 				ctx, clientset = fakekubeclient.With(ctx, tt.secret)
 			}
 			w := &Interceptor{
-				SecretGetter: interceptors.NewKubeClientSecretGetter(clientset.CoreV1(), 1024, 5*time.Second),
+				SecretGetter: interceptors.DefaultSecretGetter(clientset.CoreV1()),
 			}
 			res := w.Process(ctx, req)
 			if res.Continue {
@@ -286,7 +285,7 @@ func TestInterceptor_Process_InvalidParams(t *testing.T) {
 	ctx, _ := test.SetupFakeContext(t)
 
 	w := &Interceptor{
-		SecretGetter: interceptors.NewKubeClientSecretGetter(fakekubeclient.Get(ctx).CoreV1(), 1024, 5*time.Second),
+		SecretGetter: interceptors.DefaultSecretGetter(fakekubeclient.Get(ctx).CoreV1()),
 	}
 
 	req := &triggersv1.InterceptorRequest{
