+++ pkg/interceptors/interceptors.go
@@ -23,14 +23,10 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
-	"time"
 
 	triggersv1alpha1 "github.com/tektoncd/triggers/pkg/apis/triggers/v1alpha1"
 	triggersv1beta1 "github.com/tektoncd/triggers/pkg/apis/triggers/v1beta1"
 	"google.golang.org/grpc/codes"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/cache"
-	v1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	"knative.dev/pkg/apis"
 )
 
@@ -41,56 +37,6 @@ const (
 // Interceptor is the interface that all interceptors implement.
 type Interceptor interface {
 	ExecuteTrigger(req *http.Request) (*http.Response, error)
-}
-
-type SecretGetter interface {
-	Get(ctx context.Context, triggerNS string, sr *triggersv1beta1.SecretRef) ([]byte, error)
-}
-
-type kubeclientSecretGetter struct {
-	getter v1.SecretsGetter
-	cache  *cache.LRUExpireCache
-	ttl    time.Duration
-}
-
-type cacheKey struct {
-	triggerNS string
-	sr        triggersv1beta1.SecretRef
-}
-
-func NewKubeClientSecretGetter(getter v1.SecretsGetter, cacheSize int, ttl time.Duration) SecretGetter {
-	return &kubeclientSecretGetter{
-		getter: getter,
-		cache:  cache.NewLRUExpireCache(cacheSize),
-		ttl:    ttl,
-	}
-}
-
-// Get queries Kubernetes for the given secret reference. We use this function
-// to resolve secret material like GitHub webhook secrets, and call it once for every
-// trigger that references it.
-//
-// As we may have many triggers that all use the same secret, we cache the secret values
-// in the request cache.
-func (g *kubeclientSecretGetter) Get(ctx context.Context, triggerNS string, sr *triggersv1beta1.SecretRef) ([]byte, error) {
-	key := cacheKey{
-		triggerNS: triggerNS,
-		sr:        *sr,
-	}
-	val, ok := g.cache.Get(key)
-	if ok {
-		return val.([]byte), nil
-	}
-	secret, err := g.getter.Secrets(triggerNS).Get(ctx, sr.SecretName, metav1.GetOptions{})
-	if err != nil {
-		return nil, err
-	}
-	secretValue, ok := secret.Data[sr.SecretKey]
-	if !ok {
-		return nil, fmt.Errorf("cannot find %s key in secret %s/%s", sr.SecretKey, triggerNS, sr.SecretName)
-	}
-	g.cache.Add(key, secretValue, g.ttl)
-	return secretValue, nil
 }
 
 // GetInterceptorParams returns InterceptorParams for the current interceptors
